Fixing serious error in MonadicDeferredChain implementation.
The ".done()" call specifies the callback(s) to be called when the deferred
object is *resolved* - that is to say, when it succeeds.  In this case the
callback should be set with .always(), since it is a stub function that
determines if there was an error, and then dispatches the true success/error
callback.